Assignment 7/assignment-7-muhammed786fiyas/requirements.txt
# Core ML - CPU only
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.0+cpu
torchvision==0.16.0+cpu

# MLflow
mlflow==2.14.0
flask>=3.0.0
flask-cors<7

# Web client
streamlit>=1.32.0
requests>=2.31.0

# Data / utils
numpy>=1.24.0,<2.0
Pillow>=10.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0